--- a/napari/_vispy/layers/vectors.py
+++ b/napari/_vispy/layers/vectors.py
@@ -4,6 +4,7 @@ import numpy as np
 
 from ..visuals.vectors import VectorsVisual
 from .base import VispyBaseLayer
+from napari.layers.utils.layer_utils import segment_normal
 
 
 class VispyVectorsLayer(VispyBaseLayer):
@@ -129,45 +130,3 @@ def generate_vector_meshes_2D(vectors, width, length, p=(0, 0, 1)):
     ).astype(np.uint32)
 
     return vertices, triangles
-
-
-def segment_normal(a, b, p=(0, 0, 1)):
-    """Determines the unit normal of the vector from a to b.
-
-    Parameters
-    ----------
-    a : np.ndarray
-        Length 2 array of first point or Nx2 array of points
-    b : np.ndarray
-        Length 2 array of second point or Nx2 array of points
-    p : 3-tuple, optional
-        orthogonal vector for segment calculation in 3D.
-
-    Returns
-    -------
-    unit_norm : np.ndarray
-        Length the unit normal of the vector from a to b. If a == b,
-        then returns [0, 0] or Nx2 array of vectors
-    """
-    d = b - a
-
-    if d.ndim == 1:
-        if len(d) == 2:
-            normal = np.array([d[1], -d[0]])
-        else:
-            normal = np.cross(d, p)
-        norm = np.linalg.norm(normal)
-        if norm == 0:
-            norm = 1
-    else:
-        if d.shape[1] == 2:
-            normal = np.stack([d[:, 1], -d[:, 0]], axis=0).transpose(1, 0)
-        else:
-            normal = np.cross(d, p)
-
-        norm = np.linalg.norm(normal, axis=1, keepdims=True)
-        ind = norm == 0
-        norm[ind] = 1
-    unit_norm = normal / norm
-
-    return unit_norm