--- a/napari/_vispy/layers/vectors.py
+++ b/napari/_vispy/layers/vectors.py
@@ -2,9 +2,9 @@ from copy import copy
 
 import numpy as np
 
+from ...layers.utils.layer_utils import segment_normal
 from ..visuals.vectors import VectorsVisual
 from .base import VispyBaseLayer
-from napari.layers.utils.layer_utils import segment_normal
 
 
 class VispyVectorsLayer(VispyBaseLayer):