@sofroniewn 2022-08-01
Yes this is perfect from a git standpoint. Benchmarks looks great too. Do you want to keep making any additional changes in this PR or should I just merge?