@kcpevey 2022-08-01
@sofroniewn I ran the whole test suite and didn't hit any issues with `segment_normal`. Can you help me reproduce?