@sofroniewn 2022-08-02
> @sofroniewn I ran the whole test suite and didn't hit any issues with segment_normal. Can you help me reproduce?

it could be because you have an absolute import instead of relative - otherwise progress is looking great!!!