@sofroniewn 2022-08-02
> @sofroniewn all tests pass with this relative import. Not sure what else it may have been?

huh funny - i just tried adding import and it was ok too, maybe something else funny was going on. thanks for looking into it - i think we can just ignore and see if anyone else notices it in testing.

let me know when this PR is ready to merge, i think just a few minor tweaks to tests left - but let me know if you don't think they are needed. thanks again!!