@kcpevey 2022-08-03
I rechecked the benchmarks against main (where this branch started). 

**main**
```
[ 75.00%] ··· ...iewer_vectors.QtViewerViewVectorSuite.time_vectors_multi_refresh                 
[ 75.00%] ··· ======== =============
               param1               
              -------- -------------
                 16     9.06±0.09ms 
                 64       15.9±1ms  
                256      37.5±0.8ms 
                1024      127±2ms   
                4096      275±2ms   
               16384      430±30ms  
               65536      972±4ms   
              ======== =============

[100.00%] ··· ...k_qt_viewer_vectors.QtViewerViewVectorSuite.time_vectors_refresh                
[100.00%] ··· ======== =============
               param1               
              -------- -------------
                 16     3.03±0.04ms 
                 64      5.46±0.1ms 
                256     12.2±0.07ms 
                1024      43.6±2ms  
                4096      90.6±1ms  
               16384      138±5ms   
               65536      319±4ms   
              ======== =============
```             

**this branch**
```
[ 75.00%] ··· ...iewer_vectors.QtViewerViewVectorSuite.time_vectors_multi_refresh                
[ 75.00%] ··· ======== ============
               param1              
              -------- ------------
                 16     9.37±0.2ms 
                 64     15.6±0.2ms 
                256     38.2±0.9ms 
                1024     130±2ms   
                4096     291±10ms  
               16384     438±6ms   
               65536    1.13±0.01s 
              ======== ============

[100.00%] ··· ...k_qt_viewer_vectors.QtViewerViewVectorSuite.time_vectors_refresh                 
[100.00%] ··· ======== =============
               param1               
              -------- -------------
                 16     3.06±0.03ms 
                 64     5.38±0.07ms 
                256      12.9±0.5ms 
                1024      43.4±2ms  
                4096      95.6±5ms  
               16384     152±0.9ms  
               65536      368±4ms   
              ======== =============
```